// src/enums/ValidationMessag.enum.ts
export enum ValidationMessageEnum {
  BURGER_READY_TO_SAVE = 'burger ready to save',
  BAD_INGREDIENT = 'bad ingredient',
  OK = 'ok',
  INCORRECT_FIRST_ELEMENT = 'incorrect-first-element',
  LAST_ITEM_MUST_BE_TOP_BUN = 'last item must be top bun',
  YOU_CANNOT_REMOVE = 'You cannot remove',
  ELEMENT_NOT_EXIST = 'element not exist',
  YOU_CAN_NOT_ADD_MORE_TOP_BUN = 'you can not add more top bun',
  YOU_CAN_NOT_ADD_MORE_INGREDIENTS = 'you can not add more ingredients',
  YOUR_BURGER_IS_EMPTY = 'your burger is empty',
  YOU_CAN_NOT_REMOVE_FROM_INSIDE = 'you can not remove something from inside',
}